Changes to be committed:
	modified:   Core/CMakeLists.txt
	new file:   Core/TomoP2DSinoNum_core.c
	new file:   Core/TomoP2DSinoNum_core.h
	modified:   Core/TomoP3DModel_core.c
	modified:   Core/TomoP3DModel_core.h
	modified:   Core/utils.c
	modified:   README.md
	new file:   Wrappers/MATLAB/src/TomoP2DSinoNum.c
	modified:   Wrappers/Python/Demos/DemoModel.py
	new file:   Wrappers/Python/Demos/DemoReconASTRA.py
	modified:   Wrappers/Python/src/TomoP2D.pyx
	modified:   Wrappers/Python/src/TomoP3D.pyx
	modified:   functions/TomoP2DModelSino_core.c
	modified:   functions/TomoP2DModelSino_core.h
	modified:   functions/TomoP2DModel_core.c
	modified:   functions/TomoP2DModel_core.h
	new file:   functions/TomoP2DSinoNum.c
	new file:   functions/TomoP2DSinoNum_core.c
	new file:   functions/TomoP2DSinoNum_core.h
	modified:   functions/TomoP3DModel.c
	modified:   functions/TomoP3DModel_core.c
	modified:   functions/TomoP3DModel_core.h
	modified:   functions/TomoP3DObject.c
	modified:   functions/utils.c
	modified:   functions/utils.h
	modified:   matlab/Model2DGeneratorDemo.m
	modified:   matlab/Model3DGeneratorDemo.m
	modified:   matlab/Model3D_tGeneratorDemo.m
	modified:   matlab/Object3DGeneratorDemo.m
	modified:   matlab/install/compile_mex_linux.m
	modified:   matlab/install/compile_mex_windows.m
	modified:   python/Demos/DemoModel.py
	modified:   python/Demos/DemoModel2.py
	new file:   python/Demos/DemoModel3D.py
	new file:   python/Demos/DemoModel4D.py
	new file:   python/Demos/DemoReconASTRA.py
	renamed:    python/Demos/Reconstruction_Demo.py -> python/Demos/DemoReconTomoPy.py
	modified:   python/Demos/libraryToDict.py
	modified:   python/conda-recipe/meta.yaml
	modified:   python/setup.py
	modified:   python/src/TomoP2D.pyx
	modified:   python/src/TomoP3D.pyx
	modified:   python/tomophantom/supp/artifacts.py
	new file:   python/tomophantom/supp/recMod.py

+++ functions/TomoP2DModelSino_core.c
@@ -38,9 +38,9 @@
  */
 
 /* function to build a single sinogram - object */
-float TomoP2DObjectSino_core(float *A, int N, int P, float *Th, int AngTot, int CenTypeIn, char *Object, float C0, float x0, float y0, float a, float b, float phi_rot, int tt)
+float TomoP2DObjectSino_core(float *A, long N, long P, float *Th, long AngTot, int CenTypeIn, char *Object, float C0, float x0, float y0, float a, float b, float phi_rot, long tt)
 {
-    int i, j;
+    long i, j;
     float *Tomorange_X_Ar=NULL, Tomorange_Xmin, Tomorange_Xmax, Sinorange_Pmax, Sinorange_Pmin, H_p, H_x, C1, a22, b22, phi_rot_radian;
     float *Sinorange_P_Ar=NULL, *AnglesRad=NULL;
     float AA5, sin_2, cos_2, delta1, delta_sq, first_dr, AA2, AA3, AA6, under_exp, x00, y00;
@@ -373,7 +373,7 @@ float TomoP2DModelSino_core(float *A, int ModelSelected, int N, int P, float *Th
                                     //mexErrMsgTxt("b (object size) must be positive in [0,2] range");
                                     break; }
                                 //printf("\nObject : %s \nC0 : %f \nx0 : %f \ny0 : %f \na : %f \nb : %f \n", tmpstr2, C0, x0, y0,  a, b);
-                               TomoP2DObjectSino_core(A, N, P, Th, AngTot, CenTypeIn, tmpstr2, C0, x0, y0, b, a, -psi_gr1, 0);
+                               TomoP2DObjectSino_core(A, (long)N, (long)P, Th, (long)AngTot, CenTypeIn, tmpstr2, C0, x0, y0, b, a, -psi_gr1, 0l);
                             }
                         }
                         else {
@@ -433,14 +433,14 @@ float TomoP2DModelSino_core(float *A, int ModelSelected, int N, int P, float *Th
                                 float b_step = (b1 - b)/(steps-1);
                                 float phi_rot_step = (psi_gr1_1 - psi_gr1)/(steps-1);
                                 
-                                int tt;
+                                long tt;
                                 float x_t, y_t, a_t, b_t, C_t, phi_t, d_step;
                                 /* initialize */
                                 x_t = x0; y_t = y0; a_t = a; b_t = b; C_t = C0; phi_t = psi_gr1; d_step = d_dist;
                                 /*loop over time frames*/
                                 for(tt=0; tt < steps; tt++) {
                                     
-                                    TomoP2DObjectSino_core(A, N, P, Th, AngTot, CenTypeIn, tmpstr2, C_t, -y_t, x_t, b_t, a_t, -phi_t, tt);                                                         
+                                    TomoP2DObjectSino_core(A, (long)N, (long)P, Th, (long)AngTot, CenTypeIn, tmpstr2, C_t, -y_t, x_t, b_t, a_t, -phi_t, tt);                                                         
                                     
                                     /* calculating new coordinates of an object */
                                     if (distance != 0.0f) {
